+++ Mark/Document/Speakers/SpeakerColor.swift
@@ -1,0 +1,52 @@
+//
+//  SpeakerColor.swift
+//  Mark
+//
+//  Created by Mikhail Korzh on 13.09.2026.
+//  Copyright © 2026 Mikhail Korzh.
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with this program. If not, see <https://www.gnu.org/licenses/>.
+//
+
+import SwiftUI
+
+enum SpeakerColor: String, Codable, Sendable, CaseIterable {
+    case blue, purple, pink, red, orange, yellow, green, graphite
+
+    var color: Color {
+        switch self {
+        case .blue: .blue
+        case .purple: .purple
+        case .pink: .pink
+        case .red: .red
+        case .orange: .orange
+        case .yellow: .yellow
+        case .green: .green
+        case .graphite: .gray
+        }
+    }
+    
+    var nsColor: NSColor {
+        switch self {
+        case .blue: .systemBlue
+        case .purple: .systemPurple
+        case .pink: .systemPink
+        case .red: .systemRed
+        case .orange: .systemOrange
+        case .yellow: .systemYellow
+        case .green: .systemGreen
+        case .graphite: .systemGray
+        }
+    }
+}